In the README.md file, the changes include:
Addition of a section for profiling the code using cProfile. This section includes instructions for running the profiler and analyzing the results.
Addition of 4 new lines at the end of the file without any content (probably for formatting or separation of sections).
In the data_processing.py file, the changes include:

Addition of several import statements for numpy, numba, pandas, datetime, sqlite3, and multiprocessing.
Addition of a new function preprocess_chunk which takes a chunk of data as input and applies some preprocessing steps.
Addition of a new function preprocess_data_parallel which uses the multiprocessing library to apply the preprocess_chunk function in parallel to multiple chunks of data.
Modification of the existing preprocess_data function to include additional preprocessing steps such as calculating moving averages, momentum, volatility, RSI, and MACD.
Addition of a new if __name__ == "__main__": block at the end of the file with an example usage of the new functions.
In the utils.py file, the changes include:

Addition of import statements for functools and numba.
Modification of the existing calculate_rsi function to use numba for faster computation.
Addition of a new calculate_macd function which calculates the MACD (Moving Average Convergence Divergence) indicator for a given series. This function is also decorated with numba for faster computation.